# Improve scaling speed and some cleanup