--- a/core/trino-main/src/main/java/io/trino/SystemSessionProperties.java
+++ b/core/trino-main/src/main/java/io/trino/SystemSessionProperties.java
@@ -36,6 +36,7 @@ import java.util.Optional;
 import java.util.OptionalInt;
 
 import static com.google.common.base.Preconditions.checkArgument;
+import static io.airlift.units.DataSize.Unit.MEGABYTE;
 import static io.trino.execution.QueryManagerConfig.FAULT_TOLERANT_EXECUTION_MAX_PARTITION_COUNT_LIMIT;
 import static io.trino.execution.QueryManagerConfig.MAX_TASK_RETRY_ATTEMPTS;
 import static io.trino.plugin.base.session.PropertyMetadataUtil.dataSizeProperty;
@@ -85,6 +86,7 @@ public final class SystemSessionProperties
     public static final String MAX_WRITER_TASKS_COUNT = "max_writer_tasks_count";
     public static final String TASK_SCALE_WRITERS_MAX_WRITER_COUNT = "task_scale_writers_max_writer_count";
     public static final String WRITER_SCALING_MIN_DATA_PROCESSED = "writer_scaling_min_data_processed";
+    public static final String SKEWED_PARTITION_MIN_DATA_PROCESSED_REBALANCE_THRESHOLD = "skewed_partition_min_data_processed_rebalance_threshold";
     public static final String PUSH_TABLE_WRITE_THROUGH_UNION = "push_table_write_through_union";
     public static final String EXECUTION_POLICY = "execution_policy";
     public static final String DICTIONARY_AGGREGATION = "dictionary_aggregation";
@@ -338,6 +340,11 @@ public final class SystemSessionProperties
                         "Minimum amount of uncompressed output data processed by writers before writer scaling can happen",
                         featuresConfig.getWriterScalingMinDataProcessed(),
                         false),
+                dataSizeProperty(
+                        SKEWED_PARTITION_MIN_DATA_PROCESSED_REBALANCE_THRESHOLD,
+                        "Minimum data processed to trigger skewed partition rebalancing in local and remote exchange",
+                        DataSize.of(50, MEGABYTE),
+                        true),
                 booleanProperty(
                         PUSH_TABLE_WRITE_THROUGH_UNION,
                         "Parallelize writes when using UNION ALL in queries that write data",
@@ -1152,6 +1159,11 @@ public final class SystemSessionProperties
     public static DataSize getWriterScalingMinDataProcessed(Session session)
     {
         return session.getSystemProperty(WRITER_SCALING_MIN_DATA_PROCESSED, DataSize.class);
+    }
+
+    public static DataSize getSkewedPartitionMinDataProcessedRebalanceThreshold(Session session)
+    {
+        return session.getSystemProperty(SKEWED_PARTITION_MIN_DATA_PROCESSED_REBALANCE_THRESHOLD, DataSize.class);
     }
 
     public static boolean isPushTableWriteThroughUnion(Session session)