--- a/core/trino-main/src/main/java/io/trino/operator/exchange/LocalExchangeSourceOperator.java
+++ b/core/trino-main/src/main/java/io/trino/operator/exchange/LocalExchangeSourceOperator.java
@@ -48,7 +48,7 @@ public class LocalExchangeSourceOperator
             checkState(!closed, "Factory is already closed");
 
             OperatorContext operatorContext = driverContext.addOperatorContext(operatorId, planNodeId, LocalExchangeSourceOperator.class.getSimpleName());
-            return new LocalExchangeSourceOperator(operatorContext, localExchange.getNextSource(driverContext::getPhysicalWrittenDataSize));
+            return new LocalExchangeSourceOperator(operatorContext, localExchange.getNextSource());
         }
 
         @Override