… core/trino-main/src/main/java/io/trino/operator/exchange/LocalMergeSourceOperator.java
… core/trino-main/src/main/java/io/trino/operator/exchange/ScaleWriterExchanger.java
… core/trino-main/src/main/java/io/trino/operator/exchange/ScaleWriterPartitioningExchanger.java
… core/trino-main/src/main/java/io/trino/operator/exchange/UniformPartitionRebalancer.java
… core/trino-main/src/main/java/io/trino/operator/output/SkewedPartitionRebalancer.java
… core/trino-main/src/main/java/io/trino/sql/planner/LocalExecutionPlanner.java
… core/trino-main/src/main/java/io/trino/sql/planner/LogicalPlanner.java
… core/trino-main/src/main/java/io/trino/sql/planner/optimizations/AddExchanges.java
… core/trino-main/src/main/java/io/trino/sql/planner/optimizations/AddLocalExchanges.java
… core/trino-main/src/main/java/io/trino/sql/planner/optimizations/BeginTableWrite.java
… core/trino-main/src/main/java/io/trino/sql/planner/plan/TableWriterNode.java
… core/trino-main/src/main/java/io/trino/sql/planner/sanity/ValidateScaledWritersUsage.java
… core/trino-main/src/test/java/io/trino/connector/MockConnector.java
… core/trino-main/src/test/java/io/trino/connector/MockConnectorFactory.java
… core/trino-main/src/test/java/io/trino/execution/MockRemoteTaskFactory.java
… core/trino-main/src/test/java/io/trino/execution/TestTaskManagerConfig.java
… core/trino-main/src/test/java/io/trino/execution/TestingRemoteTaskFactory.java
… core/trino-main/src/test/java/io/trino/execution/scheduler/TestScaledWriterScheduler.java
… core/trino-main/src/test/java/io/trino/memory/TestLeastWastedEffortTaskLowMemoryKiller.java
… core/trino-main/src/test/java/io/trino/memory/TestTotalReservationOnBlockedNodesTaskLowMemoryKiller.java
… core/trino-main/src/test/java/io/trino/operator/TestTableWriterOperator.java
… core/trino-main/src/test/java/io/trino/operator/exchange/TestLocalExchange.java
… core/trino-main/src/test/java/io/trino/operator/exchange/TestUniformPartitionRebalancer.java
… core/trino-main/src/test/java/io/trino/operator/output/TestSkewedPartitionRebalancer.java
… core/trino-main/src/test/java/io/trino/server/remotetask/TestHttpRemoteTask.java
… core/trino-main/src/test/java/io/trino/sql/analyzer/TestFeaturesConfig.java
… core/trino-main/src/test/java/io/trino/sql/planner/TestingWriterTarget.java
… core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/TestRemoveEmptyMergeWriterRuleSet.java
… core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/test/PlanBuilder.java
… core/trino-main/src/test/java/io/trino/sql/planner/optimizations/TestAddExchangesScaledWriters.java