<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description


## Benchmarks

For unpartitioned case:

Before:
```
trino:insert_demo> insert into lineitem_test_unpart select * from lineitem_shipdate;
INSERT: 6733485888 rows

Query 20230620_050123_00012_e44xz, FINISHED, 7 nodes
Splits: 7,486 total, 7,486 done (100.00%)
7:24 [6.73B rows, 182GB] [15.2M rows/s, 421MB/s]
```

After:
```
trino:insert_demo> insert into lineitem_test_unpart select * from lineitem_shipdate;
INSERT: 6733485888 rows

Query 20230620_050758_00006_8c6sa, FINISHED, 7 nodes
Splits: 7,656 total, 7,656 done (100.00%)
4:31 [6.73B rows, 182GB] [24.8M rows/s, 689MB/s]
```

For partitioned case:

Before:
```
trino:insert_demo> insert into lineitem_part_target_shipmode select * from lineitem_part_3_skewed_partitions;
INSERT: 3085776220 rows

Query 20230620_052319_00017_e44xz, FINISHED, 7 nodes
Splits: 2,243 total, 2,243 done (100.00%)
3:43 [3.09B rows, 83.3GB] [13.8M rows/s, 382MB/s]
```

After:
```
trino:insert_demo> insert into lineitem_part_target_shipmode select * from lineitem_part_3_skewed_partitions;
INSERT: 3085776220 rows

Query 20230624_232648_00006_z65ur, FINISHED, 7 nodes
Splits: 2,242 total, 2,242 done (100.00%)
2:16 [3.09B rows, 83.3GB] [22.8M rows/s, 629MB/s]
```


<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or docs only and no release notes are required.
( ) Release notes are required, please propose a release note for me.
( ) Release notes are required, with the following suggested text:

```markdown
# Section
* Fix some things. ({issue}`issuenumber`)
```