@piotrolchawa-f3 2023-07-19
Hey @gaurav8297! Nice one! I was wondering about the benchmarks that you've conducted.

1. What is the Trino setup? What's the catalogue you use?
2. How does the table look like?
3. What size of servers do you use?
4. What connector do you use?

I'm asking because these tests seem great for me even before the change. We're trying to insert with benthos, and Trino chokes on 1000 rows/sec. Ofc our coordinator is kind of small (16vcpus, 32Gb ram), but I'd expect it to be able to ingest more.