@piotrolchawa-f3 2023-07-19
Hey @sopel39 , thanks for the quick response!

Our setup is:
1. Benthos ingests events from a message queue and `INSERT`s them in micro batches of 1000 using Trino.
2. Trino is using `blackhole` connector (to eliminate storage performance issues. But the target connector is Iceberg, on which we see exactly same performance).
3. Coodinator is running on 16vcpu, 32GB ram, there are two workers (we tried increasing number of workers, it changes nothing).
4. The query for insertion is: `INSERT INTO iceberg.test.my_table(id, date_0, tenant_id, comment_1, comment_2, is_t, code, is_s, date_1, round, date_2) VALUES (X, Y, Z....), (X1, Y1, Z1, ....)` We have a single insert query with 1000 rows.
5. Example row insertion: `INSERT INTO iceberg.test.my_table (id, date_0, tenant_id, comment_1, comment_2, is_t, code, is_s, date_1, round, date_2)
VALUES ('dm4X', from_iso8601_timestamp('2023-06-27'), CAST ('fa51f2df-6645-47fd-9209-198a045fbe9a' AS UUID), '11o', '15072', CAST (false AS BOOLEAN), 'EyGuVi9n', CAST (true AS BOOLEAN), from_iso8601_timestamp('2023-06-27'), TRY_CAST(NULL AS INT), from_iso8601_timestamp('2023-06-27'));`
6. Removing CASTs sped things a little, but not in any significant way.

So in our investigation we've discovered that PLANNING phase takes a long time:

```
  "queryStats" : {
    "createTime" : "2023-07-19T09:56:34.280Z",
    "executionStartTime" : "2023-07-19T09:56:34.388Z",
    "lastHeartbeat" : "2023-07-19T09:56:38.810Z",
    "endTime" : "2023-07-19T09:56:38.809Z",
    "elapsedTime" : "4.53s",
    "queuedTime" : "183.44us",
    "resourceWaitingTime" : "107.82ms",
    "dispatchingTime" : "49.48us",
    "executionTime" : "4.42s",
    "analysisTime" : "107.77ms",
    "planningTime" : "3.77s",
    "planningCpuTime" : "3.72s",
    "finishingTime" : "356.31us",
    "totalTasks" : 4,
    "runningTasks" : 0,
    "completedTasks" : 4,
    "failedTasks" : 0,
```
And if we zoom in on the PLANNER:

```
    "optimizerRulesSummaries" : [ {
      "rule" : "io.trino.sql.planner.iterative.rule.ExpressionRewriteRuleSet.ValuesExpressionRewrite",
      "invocations" : 104,
      "applied" : 2,
      "totalTime" : 3692627565,
      "failures" : 0
    }
```

which is 3.69s.


On the other hand these are the results of the query you've run for this PR:

```
trino:test> insert into table2 select * from table1;
INSERT: 22721776 rows

Query 20230719_093535_00048_afzcf, FINISHED, 3 nodes
Splits: 34 total, 34 done (100.00%)
18.04 [22.7M rows, 600MB] [1.26M rows/s, 33.3MB/s]
```
 
 Which looks fine, so probably it's for different reasons. Do you want us to create a separate issue for this?