@sopel39 2023-08-09
> Which looks fine, so probably it's for different reasons. Do you want me to create a separate issue for this?

Please create an issue for this